[Prover] Add config directory, and relocate config files.
