+++ prover-service/src/config/mod.rs
@@ -1,0 +1,4 @@
+// Copyright (c) Aptos Foundation
+
+pub mod keyless_config;
+pub mod prover_config;
