+++ prover-service/src/lib.rs
@@ -3,15 +3,14 @@
 extern crate core;
 
 pub mod api;
+pub mod config;
 pub mod deployment_information;
 pub mod error;
 pub mod groth16_vk;
 pub mod handlers;
 pub mod input_processing;
 pub mod jwk_fetching;
-pub mod keyless_config;
 pub mod metrics;
-pub mod prover_config;
 pub mod prover_state;
 pub mod training_wheels;
 pub mod utils;
