+++ prover-service/src/tests/common/types.rs
@@ -5,7 +5,7 @@ use std::process::Command;
 use std::time::{SystemTime, UNIX_EPOCH};
 
 use super::{gen_test_ephemeral_pk, gen_test_ephemeral_pk_blinder, get_test_pepper};
-use crate::prover_config::ProverServiceConfig;
+use crate::config::prover_config::ProverServiceConfig;
 use crate::tests::common::get_config;
 use crate::{
     api::{EphemeralPublicKeyBlinder, RequestInput},
