+++ prover-service/src/training_wheels/mod.rs
@@ -4,10 +4,10 @@ mod sign;
 pub mod verification_logic;
 
 use crate::api::RequestInput;
+use crate::config::prover_config::ProverServiceConfig;
 use crate::input_processing::types::VerifiedInput;
 use crate::jwk_fetching;
 use crate::jwk_fetching::get_federated_jwk;
-use crate::prover_config::ProverServiceConfig;
 use crate::prover_state::ProverServiceState;
 use crate::training_wheels::verification_logic::compute_nonce;
 use anyhow::{anyhow, bail, ensure};
